+++ CHANGELOG.md
@@ -28,7 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
     - `AUC`
 - `Multinomial` to `proba`
 - `HedgeRegressor` and `StackingBinaryClassifier` to `ensemble`
-- `QuantileLoss` to `optim`
+- `QuantileLoss` and `MiniBatcher` to `optim`
 
 ### Modified